+++ docs/CHANGELOG.md
@@ -1,6 +1,16 @@
 CHANGELOG
 ---------
 
+#### Next release
+- Bug fixes:
+  - Properly set Kotlin JVM target to Java 11
+- Update dependencies:
+  - Gradle to 7.3.1
+  - Android build tool to 32.0.0, compile and target SDK level to 32
+  - AndroidX constraint layout to 2.1.2
+  - Firebase BOM to 29.0.2, Crashlytics Gradle plugin to 2.8.1
+  - Hilt to 2.40.5
+
 #### v0.19.1 (2021-12-10)
 - Bug fixes:
   - Should also remove caches when removing translations